requirements.txt
fastapi
httpx
uvicorn[standard]
starlette
openai
python-dotenv
pydantic[email]
python-multipart
anyio

# Database
asyncpg
pgvector

# Azure Storage
azure-storage-blob

# Email
sendgrid

# Testing
pytest
pytest-asyncio
pytest-mock
pytest-cov

# MCP (keep for HTTP MCP client functionality)
mcp
langchain-mcp-adapters
langchain-core

# Microsoft Agent Framework (replaces Langgraph/Langchain)
agent-framework
azure-identity

# Document processing
docling

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-instrumentation-httpx
python-json-logger

# Circuit breaker
aiobreaker

# Microsoft Agent Framework
agent-framework